// frontend/src/types/api.ts
export enum SecurityScope {
    MOD_001_001 = "MOD_001_001",
    MOD_001_002 = "MOD_001_002",
    MOD_001_003 = "MOD_001_003",
    MOD_001_004 = "MOD_001_004",
    MOD_001_104 = "MOD_001_104",
    MOD_001_005 = "MOD_001_005",
    MOD_001_006 = "MOD_001_006",
    MOD_001_007 = "MOD_001_007",
    MOD_002_001 = "MOD_002_001",
    MOD_002_002 = "MOD_002_002",
    MOD_002_003 = "MOD_002_003",
    MOD_002_004 = "MOD_002_004",
    MOD_002_005 = "MOD_002_005",
    MOD_003_001 = "MOD_003_001",
    MOD_003_002 = "MOD_003_002",
    MOD_003_003 = "MOD_003_003",
    MOD_003_004 = "MOD_003_004",
    MOD_003_005 = "MOD_003_005",
    MOD_003_006 = "MOD_003_006",
    MOD_003_007 = "MOD_003_007",
    MOD_003_008 = "MOD_003_008",
    MOD_003_009 = "MOD_003_009",
    MOD_003_010 = "MOD_003_010",
    MOD_004_001 = "MOD_004_001",
    MOD_004_002 = "MOD_004_002",
    MOD_004_003 = "MOD_004_003",
    MOD_004_004 = "MOD_004_004",
    MOD_004_005 = "MOD_004_005",
    MOD_004_006 = "MOD_004_006",
    MOD_004_007 = "MOD_004_007",
    MOD_004_008 = "MOD_004_008",
    MOD_004_009 = "MOD_004_009",
    MOD_004_010 = "MOD_004_010",
    MOD_004_011 = "MOD_004_011"
}